requirements.txt
streamlit
faiss-cpu
pypdf
sentence-transformers
langchain
langchain-community
langchain-text-splitters
google-generativeai
